requirements.txt
face_recognition
Pillow
vk_api